refactor: added print-config.js file
